requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
pyyaml==6.0.1
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
gunicorn==23.0.0